Make ByteBuffer access RustBuffer content directly
- Rewrote JVM ByteBuffer using NIO.
- Rewrote native ByteBuffer from scratch using pointer operations.
- Made RustBuffer.asByteBuffer not copy the content as in upstream, since both implementations can access the data via pointers directly.
- Made RustBuffer.asByteBuffer check if a RustBuffer is longer than Int.MAX_VALUE bytes, as NIO does not support Long capacities.

+++ bindgen/src/templates/ffi/ByteArrayHelper.kt
@@ -2,7 +2,7 @@
 public object FfiConverterByteArray: FfiConverterRustBuffer<ByteArray> {
     override fun read(buf: ByteBuffer): ByteArray {
         val len = buf.getInt()
-        val byteArr = buf.get(len.toLong())
+        val byteArr = buf.get(len)
         return byteArr
     }
     override fun allocationSize(value: ByteArray): ULong {
